num_conv

Trait ExtendTarget

pub trait ExtendTarget<T>: ExtendTargetSealed<T> { }

A type that can be used with turbofish syntax in Extend::extend.

It is unlikely that you will want to use this trait directly. You are probably looking for the Extend trait.