Trait num_conv::TruncateTarget

pub trait TruncateTarget<T>: TruncateTargetSealed<T> { }

A type that can be used with turbofish syntax in Truncate::truncate.

It is unlikely that you will want to use this trait directly. You are probably looking for the Truncate trait.