Module num_conv::prelude

Anonymously import all extension traits.

This allows you to use the methods without worrying about polluting the namespace or importing them individually.

use num_conv::prelude::*;

Re-exports

pub use crate::CastSigned as _;

pub use crate::CastUnsigned as _;

pub use crate::Extend as _;

pub use crate::Truncate as _;